requirements.txt
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
git+https://github.com/al21e6/flask_lnbits.git
Flask-Mail==0.9.1
Flask-QRcode==3.1.0
Flask-WTF==1.0.1
html2text==2020.1.16
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.4.0
python-dotenv==0.21.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
